Create tool to fix lambda after zappa.
